Text Colour Issues & Project 1
Text that appears next to the image in the hero section was too similar in colour to the background of the image. Had to find an image that stands out from black and grey (used off-white).

Also added my first project to the website by adding the link to the GitHub repository and adding an image to represent that project.

Another change was changing the numbers on the education tiles. They would change position depending on the text in the tile but I wanted the number to have a fixed position in the bottom-right corner.